Add option to use tue-wpa2 credentials

+++ deploy/files/install.sh
@@ -7,7 +7,12 @@
 
 echo "Switching to Tue-guest network to update system, you will experience a momentary connection drop!"
 sudo cp /etc/wpa_supplicant/wpa_supplicant.conf /etc/wpa_supplicant/wpa_supplicant.conf.old
-sudo cp wpa_supplicant.conf.tue_guest /etc/wpa_supplicant/wpa_supplicant.conf
+if [ -f wpa_supplicant.conf.tue-wpa2 ]; then
+    echo "Using tue-wpa2 credentials"
+    sudo cp wpa_supplicant.conf.tue-wpa2 /etc/wpa_supplicant/wpa_supplicant.conf
+else
+    sudo cp wpa_supplicant.conf.tue_guest /etc/wpa_supplicant/wpa_supplicant.conf
+fi
 sudo wpa_cli -i wlan0 reconfigure
 sudo wpa_cli -i wlan0 reconnect
 sleep 1
